Eliminating warnings from SwifShader
SwiftShader had over 270 warnings and this
reduces the count to about 90, some of which
can be addressed later, since some are less
trivial to remove. The changes include:
- Making sure we don't build any project with
  RTTI enabled, otherwise, we get some
  vftable size mismatch warnings
- Making sure exception handling is the same
  for all configurations
- Changing minus <unsigned> to the equivalent,
  and valid, ~<unsigned> + 1
- Explicitly casting types to bool by adding
  "!= 0" or "!= nullptr", depending on the case
- Making sure no #define overrules another one
  without first doing the proper #undef of that
  #define
- Removing signed VS unsigned comparisons
- Making sure WinMain is __stdcall

WebGL tests passed.

Reviewed-on: https://swiftshader-review.googlesource.com/3475
Tested-by: Alexis Hétu <sugoi@google.com>

+++ src/OpenGL/compiler/ParseHelper.cpp
@@ -481,7 +481,7 @@ bool TParseContext::constructorErrorCheck(int line, TIntermNode* node, TFunction
     bool overFull = false;
     bool matrixInMatrix = false;
     bool arrayArg = false;
-    for (int i = 0; i < function.getParamCount(); ++i) {
+    for (size_t i = 0; i < function.getParamCount(); ++i) {
         const TParameter& param = function.getParam(i);
         size += param.type->getObjectSize();
